git-svn-id: http://192.168.20.11/sequoiadb/trunk@11503 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdAgent.cpp
@@ -1787,7 +1787,7 @@ namespace engine
                   probe = 80 ;
                   goto error ;
                }
-               SDB_ASSERT ( replyHeader.header.messageLength ==
+               SDB_ASSERT ( (UINT32)replyHeader.header.messageLength ==
                             sizeof(replyHeader) + pObj->objsize(),
                             "Reply size doesn't match actual" )
             }
